Distribute Random.swift & misc test cleanup
