+++ Sources/MapboxMaps/Foundation/CameraBounds.swift
