+++ Sources/MapboxMaps/Foundation/CameraBoundsOptions.swift
