+++ Sources/MapboxMaps/Foundation/CameraOptions.swift
