+++ Sources/MapboxMaps/Foundation/CameraState.swift
