+++ Tests/MapboxMapsTests/Annotations/ViewAnnotationPositionDescriptor.swift
