+++ Tests/MapboxMapsTests/Foundation/CameraOptionsTests.swift
