+++ Tests/MapboxMapsTests/Foundation/CameraStateTests.swift
