+++ Tests/MapboxMapsTests/Helpers/Random/Array+Random.swift
@@ -1,0 +1,5 @@
+extension Array {
+    static func random(withLength length: UInt, generator: () -> Element) -> Self {
+        return (0..<length).reduce(into: []) { array, _ in array.append(generator()) }
+    }
+}
