+++ Tests/MapboxMapsTests/Helpers/Random/CLAccuracyAuthorization+Random.swift
@@ -1,0 +1,7 @@
+import CoreLocation
+
+extension CLAccuracyAuthorization {
+    static func random() -> Self {
+        return .random() ? .fullAccuracy : .reducedAccuracy
+    }
+}
