+++ Tests/MapboxMapsTests/Helpers/Random/CLLocation+Random.swift
@@ -1,0 +1,9 @@
+import CoreLocation
+
+extension CLLocation {
+    static func random() -> CLLocation {
+        return CLLocation(
+            latitude: .random(in: -89...89),
+            longitude: .random(in: -180..<180))
+    }
+}
