+++ Tests/MapboxMapsTests/Helpers/Random/CLLocationCoordinate2D+Random.swift
@@ -1,0 +1,9 @@
+import CoreLocation
+
+extension CLLocationCoordinate2D {
+    static func random() -> Self {
+        return CLLocationCoordinate2D(
+            latitude: .random(in: -90...90),
+            longitude: .random(in: -180..<180))
+    }
+}
