+++ Tests/MapboxMapsTests/Helpers/Random/Character+Random.swift
@@ -1,0 +1,5 @@
+extension Character {
+    static func randomASCII() -> Self {
+        return Character(UnicodeScalar(.random(in: 0x20...0x7E))!)
+    }
+}
