+++ Tests/MapboxMapsTests/Helpers/Random/Optional+Random.swift
@@ -1,0 +1,5 @@
+extension Optional {
+    static func random(_ generator: @autoclosure () -> Wrapped) -> Self {
+        return .random() ? .none : .some(generator())
+    }
+}
