+++ Tests/MapboxMapsTests/Helpers/Random/String+Random.swift
@@ -1,0 +1,5 @@
+extension String {
+    static func randomASCII(withLength length: UInt) -> Self {
+        return (0..<length).reduce(into: "") { s, _ in s.append(.randomASCII()) }
+    }
+}
