+++ Tests/MapboxMapsTests/Helpers/XCTestCase+GuardForMetalDevice.swift
@@ -1,0 +1,9 @@
+import XCTest
+
+extension XCTestCase {
+    func guardForMetalDevice() throws {
+        guard MTLCreateSystemDefaultDevice() != nil else {
+            throw XCTSkip("No valid Metal device (OS version or VM?)")
+        }
+    }
+}
