+++ Tests/MapboxMapsTests/Integration Tests/Map/DidIdleFailureIntegrationTest.swift
