+++ Tests/MapboxMapsTests/Integration Tests/Map/MapViewRenderedSnapshotIntegrationTests.swift
