+++ Tests/MapboxMapsTests/Location/Random/Location+Random.swift
@@ -1,0 +1,10 @@
+import MapboxMaps
+
+extension Location {
+    static func random() -> Location {
+        return Location(
+            location: .random(),
+            heading: .random(MockHeading()),
+            accuracyAuthorization: .random())
+    }
+}
