+++ Tests/MapboxMapsTests/Viewport/Random/ViewportOptions+Random.swift
@@ -1,0 +1,8 @@
+@_spi(Experimental) import MapboxMaps
+
+extension ViewportOptions {
+    static func random() -> Self {
+        return ViewportOptions(
+            transitionsToIdleUponUserInteraction: .random())
+    }
+}
