+++ Tests/MapboxMapsTests/Viewport/Random/ViewportStatus+Random.swift
@@ -1,0 +1,11 @@
+@_spi(Experimental) import MapboxMaps
+
+extension ViewportStatus {
+    static func random() -> Self {
+        return [
+            .idle,
+            .state(MockViewportState()),
+            .transition(MockViewportTransition(), toState: MockViewportState())
+        ].randomElement()!
+    }
+}
