+++ Tests/MapboxMapsTests/Viewport/States/FollowPuck/Random/FollowPuckViewportStateBearing+Random.swift
@@ -1,0 +1,11 @@
+@_spi(Experimental) import MapboxMaps
+
+extension FollowPuckViewportStateBearing {
+    static func random() -> Self {
+        return [
+            .constant(.random(in: 0..<360)),
+            .heading,
+            .course
+        ].randomElement()!
+    }
+}
