+++ Tests/MapboxMapsTests/Viewport/Transitions/Default/Random/DefaultViewportTransitionOptions+Random.swift
@@ -1,0 +1,8 @@
+@_spi(Experimental) import MapboxMaps
+
+extension DefaultViewportTransitionOptions {
+    static func random() -> Self {
+        return DefaultViewportTransitionOptions(
+            maxDuration: .random(in: 0...20))
+    }
+}
